requirements.txt
asgiref==3.2.3
backcall==0.1.0
decorator==4.3.0
Django==3.0.4
django-bootstrap3==12.0.3
django-cors-headers==3.2.1
django-pure-pagination==0.3.0
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyMySQL==0.9.3
pytz==2018.4
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
captcha
django-crispy-forms
django-filter
django-pure-pagination
djangorestframework
django-rest-swagger